refactor(ENV): 将环境判断放到基础组件内部
