+++ src/components/index.ts
@@ -5,3 +5,4 @@ export * from './Btn';
 export * from './Icon';
 export * from './Modal';
 export * from './Card';
+export * from './Hoverable';
